feat: Minimum visible width/height
The main thrust of this commit is to prevent the overlay from being
dragged to the edge or off the screen, after which the user is unable to
retrieve it. This package is not in a great state and rather than trying
to pick apart the various incompatible options and weird bugs I've opted
to just hack it up to get it to a working state for our needs. The
quality of this package is poor enough that we'd be better off writing
our own in the future rather than trying to extend this one.

There were some gradle updates made and I bumped the minimum sdk version
to allow the use of a more reliable method for determining the whole
screen size.

It's not perfect, for whatever reason the detection of the screen size
and the x/y offset seem to differ on different phones and Android
verisons. However this seems good enough and it plays it on the safe
side, worst case we force a little too much of the overlay to remain
visible.

I'd love to be able to avoid the overlay code being executed on app
startup but my attempts at that were unsuccessful and that's something
we consider in a future rewrite of this package.