fix: Disable service restarting
This prevents an ANR and crash report when Android decides to kill the
service and then tries to revive it (due to START_STICKY).

I want to make the overlay resilient to restarts but that will take some
work so for now this should stop ANRs which can impact our store
listing.